# backend/src/syntara/schemas/forms/openapi.yaml
---
openapi: 3.1.0
info:
  title: Orchestrator Forms API
  version: 1.0.0
  description: |
    Form prompts API for the Orchestrator platform.

    Minimal internal-facing implementation for workflow engine integration.
    AAP-91889 will extend with full filtering/sorting/enrichment.

servers:
  - url: /api/v1
    description: API v1

tags:
  - name: Form Prompts
    description: Form prompt management for workflow human-in-the-loop gates

paths:
  /form_prompts:
    get:
      summary: List form prompts
      description: |-
        List form prompts filtered by execution ID.
        Internal endpoint for expire/cancel activities.
      operationId: list_form_prompts
      tags:
        - Form Prompts
      x-app-permission:
        resource: form_prompt
        action: read
      parameters:
        - name: execution_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
            title: Execution Id
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - $ref: "#/components/schemas/FormPromptStatus"
              - type: 'null'
            title: Status
      responses:
        "200":
          description: List of form prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FormPrompt"
        "400":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/BadRequestError"
        "401":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/UnauthorizedError"
        "403":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/ForbiddenError"
        "500":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/InternalServerError"
      security:
        - bearerAuth: []
    post:
      summary: Create form prompt
      description: |-
        Create a new form prompt.

        Internal service-to-service endpoint for workflow engine.
        Gated to admin for now; replace with proper service-to-service auth.
      operationId: create_form_prompt
      tags:
        - Form Prompts
      x-app-permission:
        resource: form_prompt
        action: create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormPromptCreateRequest"
      responses:
        "201":
          description: Form prompt created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormPrompt"
        "400":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/BadRequestError"
        "401":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/UnauthorizedError"
        "403":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/ForbiddenError"
        "409":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/ConflictError"
        "422":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/ValidationError"
        "500":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/InternalServerError"
      security:
        - bearerAuth: []

  /form_prompts/batch:
    post:
      summary: Batch update form prompt statuses
      description: |-
        Batch update form prompt statuses.
        Internal endpoint for expire/cancel activities.
      operationId: batch_update_form_prompts
      tags:
        - Form Prompts
      x-app-permission:
        resource: form_prompt
        action: create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchFormPromptRequest"
      responses:
        "200":
          description: Batch update results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUpdateResponse"
        "400":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/BadRequestError"
        "401":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/UnauthorizedError"
        "403":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/ForbiddenError"
        "422":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/ValidationError"
        "500":
          $ref: "../base/shared-resources.openapi.yaml#/components/responses/InternalServerError"
      security:
        - bearerAuth: []

components:
  schemas:
    # Form definition schemas (from original spec)
    FormDefinition:
      type: object
      title: FormDefinition
      description: Complete form definition with fields and metadata.
      additionalProperties: false
      required:
        - fields
      properties:
        fields:
          type: array
          title: Fields
          minItems: 1
          maxItems: 100
          items:
            oneOf:
              - $ref: '#/components/schemas/TextField'
              - $ref: '#/components/schemas/TextAreaField'
              - $ref: '#/components/schemas/MaskedTextField'
              - $ref: '#/components/schemas/EmailField'
              - $ref: '#/components/schemas/NumberField'
              - $ref: '#/components/schemas/CheckboxField'
              - $ref: '#/components/schemas/DateField'
              - $ref: '#/components/schemas/DropdownField'
              - $ref: '#/components/schemas/MultiSelectField'
            discriminator:
              propertyName: type
              mapping:
                text: '#/components/schemas/TextField'
                textarea: '#/components/schemas/TextAreaField'
                masked_text: '#/components/schemas/MaskedTextField'
                email: '#/components/schemas/EmailField'
                number: '#/components/schemas/NumberField'
                checkbox: '#/components/schemas/CheckboxField'
                date: '#/components/schemas/DateField'
                dropdown: '#/components/schemas/DropdownField'
                multi_select: '#/components/schemas/MultiSelectField'

    TextField:
      type: object
      title: TextField
      description: Text field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: text
          title: Type
        default:
          anyOf:
            - type: string
            - type: 'null'
          title: Default

    TextAreaField:
      type: object
      title: TextAreaField
      description: Text area field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: textarea
          title: Type
        default:
          anyOf:
            - type: string
            - type: 'null'
          title: Default

    MaskedTextField:
      type: object
      title: MaskedTextField
      description: Masked text field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: masked_text
          title: Type
        default:
          anyOf:
            - type: string
            - type: 'null'
          title: Default

    EmailField:
      type: object
      title: EmailField
      description: Email field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: email
          title: Type
        default:
          anyOf:
            - type: string
            - type: 'null'
          title: Default

    NumberField:
      type: object
      title: NumberField
      description: Numeric field supporting int or float.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: number
          title: Type
        default:
          anyOf:
            - type: number
            - type: integer
            - type: 'null'
          title: Default

    CheckboxField:
      type: object
      title: CheckboxField
      description: Boolean checkbox field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
          description: When true the checkbox must be checked to submit, for example a terms of service or acknowledgment. An unchecked required checkbox fails with error code 'must_be_checked'.
        type:
          type: string
          const: checkbox
          title: Type
        default:
          type: boolean
          title: Default
          default: false

    DateField:
      type: object
      title: DateField
      description: Date field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: date
          title: Type
        default:
          anyOf:
            - type: string
            - type: 'null'
          title: Default

    DropdownField:
      type: object
      title: DropdownField
      description: Dropdown field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
        - options
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: dropdown
          title: Type
        options:
          title: Options
          oneOf:
            - $ref: '#/components/schemas/StaticOptions'
            - $ref: '#/components/schemas/DynamicOptions'
          discriminator:
            propertyName: source
            mapping:
              static: '#/components/schemas/StaticOptions'
              dynamic: '#/components/schemas/DynamicOptions'
        default:
          anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - type: 'null'
          title: Default

    MultiSelectField:
      type: object
      title: MultiSelectField
      description: Multi-select field.
      additionalProperties: false
      required:
        - value_name
        - label
        - type
        - options
      properties:
        value_name:
          type: string
          title: Value Name
          maxLength: 64
          pattern: ^[a-zA-Z_][a-zA-Z0-9_]*$
        label:
          type: string
          title: Label
          minLength: 1
          maxLength: 200
        placeholder:
          anyOf:
            - type: string
            - type: 'null'
          title: Placeholder
        help_text:
          anyOf:
            - type: string
            - type: 'null'
          title: Help Text
        required:
          type: boolean
          title: Required
          default: false
        type:
          type: string
          const: multi_select
          title: Type
        options:
          title: Options
          oneOf:
            - $ref: '#/components/schemas/StaticOptions'
            - $ref: '#/components/schemas/DynamicOptions'
          discriminator:
            propertyName: source
            mapping:
              static: '#/components/schemas/StaticOptions'
              dynamic: '#/components/schemas/DynamicOptions'
        default:
          anyOf:
            - type: array
              items: {}
            - type: 'null'
          title: Default

    StaticOptions:
      type: object
      title: StaticOptions
      description: Static option list for dropdown or multi-select.
      additionalProperties: false
      required:
        - source
        - values
      properties:
        source:
          type: string
          const: static
          title: Source
        values:
          type: array
          title: Values
          minItems: 1
          maxItems: 500
          items:
            $ref: '#/components/schemas/StaticOption'

    StaticOption:
      type: object
      title: StaticOption
      description: A single static option for dropdown or multi-select.
      additionalProperties: false
      required:
        - display_label
        - value
      properties:
        display_label:
          type: string
          title: Display Label
          minLength: 1
          maxLength: 200
        value:
          anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
          title: Value

    DynamicOptions:
      type: object
      title: DynamicOptions
      description: Dynamic option list resolved from upstream node output.
      additionalProperties: false
      required:
        - source
        - expression
      properties:
        source:
          type: string
          const: dynamic
          title: Source
        expression:
          type: string
          title: Expression
        value_key:
          anyOf:
            - type: string
            - type: 'null'
          title: Value Key
        label_key:
          anyOf:
            - type: string
            - type: 'null'
          title: Label Key

    # Form prompt schemas (new for workflow engine integration)
    FormPrompt:
      type: object
      title: FormPrompt
      description: Form prompt basic response (minimal for workflow engine)
      required:
        - id
        - execution_id
        - project_id
        - prompt_node_id
        - name
        - status
      properties:
        id:
          type: string
          format: uuid
        execution_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        prompt_node_id:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/FormPromptStatus"
        loop_iteration_path:
          type: array
          items:
            type: integer
        temporal_activity_id:
          anyOf:
            - type: string
            - type: "null"

    FormPromptStatus:
      type: string
      title: FormPromptStatus
      enum:
        - pending
        - submitted
        - expired
        - cancelled

    FormPromptCreateRequest:
      type: object
      title: FormPromptCreateRequest
      description: Request payload for creating a form prompt (internal schema)
      required:
        - execution_id
        - project_id
        - prompt_node_id
        - name
        - form_definition
        - workflow_context
      properties:
        execution_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        prompt_node_id:
          type: string
        name:
          type: string
          maxLength: 255
          minLength: 1
        message:
          anyOf:
            - type: string
              maxLength: 2000
            - type: "null"
        loop_iteration_path:
          type: array
          items:
            type: integer
          default: []
        temporal_activity_id:
          anyOf:
            - type: string
              maxLength: 255
            - type: "null"
        timeout_at:
          anyOf:
            - type: string
              format: date-time
            - type: "null"
        form_definition:
          type: object
          description: FormDefinition object (validated server-side)
        submit_label:
          anyOf:
            - type: string
              maxLength: 64
            - type: "null"
        success_message:
          anyOf:
            - type: string
              maxLength: 500
            - type: "null"
        timezone:
          anyOf:
            - type: string
              maxLength: 64
            - type: "null"
        css_override:
          anyOf:
            - type: string
              maxLength: 10000
            - type: "null"
        workflow_context:
          type: object
          description: WorkflowContext object
        responder_user_ids:
          anyOf:
            - type: array
              maxItems: 1000
              items:
                type: string
                format: uuid
            - type: "null"
        responder_group_ids:
          anyOf:
            - type: array
              maxItems: 1000
              items:
                type: string
                format: uuid
            - type: "null"

    BatchFormPromptStatus:
      type: string
      title: BatchFormPromptStatus
      enum:
        - expired
        - cancelled

    BatchFormPromptUpdate:
      type: object
      title: BatchFormPromptUpdate
      required:
        - prompt_id
        - status
      properties:
        prompt_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/BatchFormPromptStatus"
        notes:
          anyOf:
            - type: string
              maxLength: 2000
            - type: "null"

    BatchFormPromptRequest:
      type: object
      title: BatchFormPromptRequest
      required:
        - updates
      properties:
        updates:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BatchFormPromptUpdate"

    BatchUpdateResponse:
      type: object
      title: BatchUpdateResponse
      required:
        - results
        - total_success
        - total_failed
      properties:
        results:
          type: array
          items:
            type: object
        total_success:
          type: integer
          minimum: 0
        total_failed:
          type: integer
          minimum: 0

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []